move xrepo logic from Database to Backend
