+++ lsif/src/typescript-linked-reference-results.test.ts
@@ -55,14 +55,11 @@ describe('Database', () => {
         }
 
         return new Database(
-            storageRoot,
-            xrepoDatabase,
             connectionCache,
             documentCache,
             resultChunkCache,
             dump.id,
-            dbFilename(storageRoot, dump.id, dump.repository, dump.commit),
-            ''
+            dbFilename(storageRoot, dump.id, dump.repository, dump.commit)
         )
     }
